requirements.txt
pyyaml
click
fastapi[all]
requests
cryptography
hvac
